+++ rill-flow-service/src/main/java/com/weibo/rill/flow/service/manager/DescriptorManager.java
@@ -542,11 +542,4 @@ public class DescriptorManager {
         Optional.ofNullable(thirdPart).ifPresent(ids::add);
         return StringUtils.join(ids, ReservedConstant.COLON);
     }
-
-    /**
-     * change descriptorId to businessId
-     */
-    public static String changeDescriptorIdToBusinessId(String descriptorId) {
-        return StringUtils.substringBefore(descriptorId, ReservedConstant.COLON);
-    }
 }
