+++ rill-flow-service/src/main/java/com/weibo/rill/flow/service/util/ExecutionIdUtil.java
@@ -128,6 +128,13 @@ public class ExecutionIdUtil {
         return bucketName.replaceFirst(BUCKET_NAME_PREFIX, StringUtils.EMPTY);
     }
 
+    /**
+     * change descriptorId to businessId
+     */
+    public static String changeDescriptorIdToBusinessId(String descriptorId) {
+        return StringUtils.substringBefore(descriptorId, ReservedConstant.COLON);
+    }
+
     private ExecutionIdUtil() {
 
     }
